first front end, diffent pages and logic

+++ backend/controllers/product.controller.js
@@ -50,12 +50,19 @@ export const updateProduct = async (req, res) => {
 };
 
 export const deleteProduct = async (req, res) => {
+
+   
    const { id } = req.params;
+   
+   if (!moongoose.Types.ObjectId.isValid(id)) {
+      return res.status(404).json({ success: false, message: 'Invalid Product ID' });
+   }
+
    try {
       await Product.findByIdAndDelete(id);
       res.status(200).json({ success: true, message: 'Product deleted successfully' });
    } catch (error) {
       console.error("Error in Delete Product: ", error);
-      res.status(404).json({ success: false, message: 'Product not found' });
+      res.status(500).json({ success: false, message: 'Server error' });
    }
 };
